pyproject.toml
[project]
name = "mk2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0.1",
    "dotenv>=0.9.9",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"


[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
